Upgraded /trinsfirmir command
Added the "reply to a text" message and modified the code where the vowels are replaced to make it shorter.
Note: I had to separate the normal vowels with tilde vowels because it gives an "string out of index" error. besides that the code works very well with no errors so far.