Refactor rep command, don't error out on members that left the group
Previously, when trying to generate the reputation list, Telegram would
throw an error if we tried to link to an user that left the group. We
now separate old group users from the current group users and link only
to users existing in the group.
Also, the refactor paves the way to adding a paginated reputation list
for groups that contain many users, as previously we would just generate
a long list (that would result in errors if we had to list lots of
users, as groups can have up to 200 thousand members).