Paginate reputation list
In groups with too many members, trying to list all users in a single
message could cause an error as the message would be too long. To solve
this, we paginate the reputation list into chunks of 15 users per page.
The code had to be restructured a bit since the latest commit, as we
were paginating only taking the current group members into account,
which resulted in pages with more than 15 users.