Prevent multiple messages sent at the same time from increasing the nisman count more than once
If two different users send a message at the same time, it would still count 2 nismans. However, the chances of that happening are close to 0, and if it does, it would be pretty funny, so I'm leaving that in. If it's not desired, just storing the chat_id in checking_nsm (instead of chat_id + user) should do the job.